mailer/requirements.txt
pydantic==2.11.3
boto3==1.38.3
requests==2.32.3
python-dotenv==1.1.0
